requirements.txt
streamlit

llama-index
llama-index-vector-stores-qdrant
llama-index-llms-openai
llama-index-llms-anthropic
llama-index-llms-ollama
llama-index-embeddings-openai
llama-index-embeddings-ollama
llama-index-embeddings-fastembed
llama-index-embeddings-huggingface
qdrant-client

python-dotenv